Refactor
- Refactored deuceclient.api.vault.Vault to be more consistent with how it stores information, using the __properties dictionary, with the other objects in the API
- Added the missing blocks and files properties to the deuceclient.api.vault.Vault
- Refactored deuceclient.client.deuce to remove the __vault* functions
--- Most all functions now require the vault parameter to be the Vault object not the vault name; there are a few exceptions they should be well documented (if not, that's a bug)
--- TypeError is raised in explicit check cases
- Refactored deuceclient.client.deuce to change ProjectId property to project_id for consistency
- Refactored deuceclient.client.deuce for how to get a Vault
--- CreateVault now creates a vault and returns a Vault object
--- Added a GetVault to retrieve a vault object for a verified existing vault
- Integrated the Block listing to add the return value to the Vault's blocks property

+++ deuceclient/client/deuce.py
@@ -15,23 +15,6 @@ class DeuceClient(Command):
     """
     Object defining HTTP REST API calls for interacting with Deuce.
     """
-    @staticmethod
-    def __vault_id(vault):
-        if isinstance(vault, api_vault.Vault):
-            return vault.vault_id
-        else:
-            return vault
-
-    @staticmethod
-    def __vault_status(vault, status):
-        if isinstance(vault, api_vault.Vault):
-            vault.status = status
-
-    @staticmethod
-    def __vault_statistics(vault, stats):
-        if isinstance(vault, api_vault.Vault):
-            vault.statistics = stats
-
     def __init__(self, authenticator, apihost, sslenabled=False):
         """
         Initialize the Deuce Client access
@@ -51,7 +34,7 @@ class DeuceClient(Command):
         Update common headers
         """
         self.headers['X-Auth-Token'] = self.authenticator.AuthToken
-        self.headers['X-Project-ID'] = self.ProjectId
+        self.headers['X-Project-ID'] = self.project_id
 
     def __log_request_data(self):
         """
@@ -63,18 +46,22 @@ class DeuceClient(Command):
         self.log.debug('uri: %s', self.Uri)
 
     @property
-    def ProjectId(self):
+    def project_id(self):
         """
         Return the project id to use
         """
         return self.authenticator.AuthTenantId
 
-    def CreateVault(self, vault):
+    def CreateVault(self, vault_name):
+        """Create a vault
+
+        :param vault_name: name of the vault
+        :returns: deuceclient.api.Vault instance of the new Vault
         """
-        Create a Vault
-            vault - name of vault to be created
-        """
-        path = api_v1.get_vault_path(self.__vault_id(vault))
+        if not isinstance(vault_name, str):
+            raise TypeError('vault_name must be string')
+
+        path = api_v1.get_vault_path(vault_name)
         self.ReInit(self.sslenabled, path)
 
         self.__update_headers()
@@ -82,26 +69,49 @@ class DeuceClient(Command):
         res = requests.put(self.Uri, headers=self.Headers)
 
         if res.status_code == 201:
-            DeuceClient.__vault_status(vault, 'created')
-            return True
+            vault = api_vault.Vault(project_id=self.project_id,
+                                    vault_id=vault_name)
+            vault.status = "created"
+            return vault
         else:
             raise RuntimeError(
                 'Failed to create Vault. '
                 'Error ({0:}): {1:}'.format(res.status_code, res.text))
 
+    def GetVault(self, vault_name):
+        """Get an existing vault
+
+        :param vault_name: name of the vault
+        :returns: deuceclient.api.Vault instance of the existing Vault
+        """
+        if not isinstance(vault_name, str):
+            raise TypeError('vault_name must be string')
+
+        if self.VaultExists(vault_name):
+            vault = api_vault.Vault(project_id=self.project_id,
+                                    vault_id=vault_name)
+            vault.status = "valid"
+            return vault
+        else:
+            raise RuntimeError('Failed to find a Vault with the name {0:}'
+                               .format(vault_name))
+
     def DeleteVault(self, vault):
         """
         Delete a Vault
-            vault - name of vault to be deleted
+            vault - vault or name of the vault to be deleted
         """
-        path = api_v1.get_vault_path(self.__vault_id(vault))
+        if not isinstance(vault, api_vault.Vault):
+            raise TypeError('vault must be deuceclient.api.Vault')
+
+        path = api_v1.get_vault_path(vault.vault_id)
         self.ReInit(self.sslenabled, path)
         self.__update_headers()
         self.__log_request_data()
         res = requests.delete(self.Uri, headers=self.Headers)
 
         if res.status_code == 204:
-            DeuceClient.__vault_status(vault, 'deleted')
+            vault.status = 'deleted'
             return True
         else:
             raise RuntimeError(
@@ -113,17 +123,25 @@ class DeuceClient(Command):
         Return the statistics on a Vault
             vault - name of vault to be deleted
         """
-        path = api_v1.get_vault_path(self.__vault_id(vault))
+        # Note: We cannot use GetVault() here b/c it would
+        #   end up being self-referential
+        vault_id = vault
+        if isinstance(vault, api_vault.Vault):
+            vault_id = vault.vault_id
+
+        path = api_v1.get_vault_path(vault_id)
         self.ReInit(self.sslenabled, path)
         self.__update_headers()
         self.__log_request_data()
         res = requests.head(self.Uri, headers=self.Headers)
 
         if res.status_code == 204:
-            DeuceClient.__vault_status(vault, 'valid')
+            if isinstance(vault, api_vault.Vault):
+                vault.status = 'valid'
             return True
         elif res.status_code == 404:
-            DeuceClient.__vault_status(vault, 'invalid')
+            if isinstance(vault, api_vault.Vault):
+                vault.status = 'invalid'
             return False
         else:
             raise RuntimeError(
@@ -135,16 +153,18 @@ class DeuceClient(Command):
         Return the statistics on a Vault
             vault - name of vault to be deleted
         """
-        path = api_v1.get_vault_path(self.__vault_id(vault))
+        if not isinstance(vault, api_vault.Vault):
+            raise TypeError('vault must be of type deuceclient.api.Vault')
+
+        path = api_v1.get_vault_path(vault.vault_id)
         self.ReInit(self.sslenabled, path)
         self.__update_headers()
         self.__log_request_data()
         res = requests.get(self.Uri, headers=self.Headers)
 
         if res.status_code == 200:
-            statistics = res.json()
-            DeuceClient.__vault_statistics(vault, statistics)
-            return statistics
+            vault.statistics = res.json()
+            return vault
         else:
             raise RuntimeError(
                 'Failed to get Vault statistics. '
@@ -154,7 +174,7 @@ class DeuceClient(Command):
         """
         Return the list of blocks in the vault
         """
-        url = api_v1.get_blocks_path(self.__vault_id(vault))
+        url = api_v1.get_blocks_path(vault.vault_id)
         if marker is not None or limit is not None:
             # add the separator between the URL and the parameters
             url = url + '?'
@@ -176,7 +196,11 @@ class DeuceClient(Command):
         res = requests.get(self.Uri, headers=self.Headers)
 
         if res.status_code == 200:
-            return res.json()
+            for block_entry in res.json():
+                vault.blocks[block_entry] = api.Block(vault.project_id,
+                                                      vault.vault_id,
+                                                      block_entry)
+            return True
         else:
             raise RuntimeError(
                 'Failed to get Block list for Vault . '
@@ -190,7 +214,7 @@ class DeuceClient(Command):
                       f.e 74bdda817d796333e9fe359e283d5643ee1a1397
             blockcontent - data present in the block to uploaded
         """
-        url = api_v1.get_block_path(self.__vault_id(vault), blockid)
+        url = api_v1.get_block_path(vault.vault_id, blockid)
         self.ReInit(self.sslenabled, url)
         self.__update_headers()
         self.__log_request_data()
@@ -211,7 +235,7 @@ class DeuceClient(Command):
         Delete the block from the vault.
         This funciton has not been tested
         """
-        url = api_v1.get_block_path(self.__vault_id(vault), blockid)
+        url = api_v1.get_block_path(vault.vault_id, blockid)
         self.ReInit(self.sslenabled, url)
         self.__update_headers()
         self.__log_request_data()
@@ -229,7 +253,7 @@ class DeuceClient(Command):
         vault - exisiting vault, eg 'v1'
         block id - sha1 of block, eg - 74bdda817d796333e9fe359e283d5643ee1a1397
         """
-        url = api_v1.get_block_path(self.__vault_id(vault), blockid)
+        url = api_v1.get_block_path(vault.vault_id, blockid)
         self.ReInit(self.sslenabled, url)
         self.__update_headers()
         self.__log_request_data()
@@ -247,7 +271,7 @@ class DeuceClient(Command):
         Creates a file in the specified vault, does not post data to it
         Returns the location of the file which gives the file id
         """
-        url = api_v1.get_files_path(self.__vault_id(vault))
+        url = api_v1.get_files_path(vault.vault_id)
         self.ReInit(self.sslenabled, url)
         self.__update_headers()
         self.__log_request_data()
@@ -288,7 +312,7 @@ class DeuceClient(Command):
                 }
         Mandatory to supply block size and offset along with the block id
         """
-        url = api_v1.get_file_path(self.__vault_id(vault), fileid)
+        url = api_v1.get_file_path(vault.vault_id, fileid)
         self.ReInit(self.sslenabled, url)
         self.__update_headers()
         self.__log_request_data()
@@ -309,7 +333,7 @@ class DeuceClient(Command):
         This function is returning a 404
         """
 
-        url = api_v1.get_fileblocks_path(self.__vault_id(vault), fileid)
+        url = api_v1.get_fileblocks_path(vault.vault_id, fileid)
 
         if marker is not None or limit is not None:
             # add the separator between the URL and the parameters
